Reorganize blgo-specific components
