+++ src/components/blog/wp/config.ts
