+++ src/components/blog/wp/media.ts
